free_api/requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
pytest>=6.2.5
pandas>=1.3.0
microbilt-api>=1.0.0 